pose-data-visualisation/requirements.txt
Package                 Version
----------------------- ----------------------------------------------------------------------------------
absl-py                 2.1.0
attrs                   21.2.0
Automat                 20.2.0
bcc                     0.18.0
bcrypt                  3.2.0
blinker                 1.4
Brlapi                  0.8.3
certifi                 2020.6.20
cffi                    1.17.1
chardet                 4.0.0
chrome-gnome-shell      0.0.0
click                   8.0.3
colorama                0.4.4
command-not-found       0.3
constantly              15.1.0
contourpy               1.3.1
cryptography            3.4.8
cupshelpers             1.0
cycler                  0.12.1
dbus-python             1.2.18
defer                   1.0.6
deluge                  2.0.3
distro                  1.7.0
flatbuffers             24.12.23
fonttools               4.55.3
GeoIP                   1.3.2
hidpidaemon             18.4.6
httplib2                0.20.2
hyperlink               21.0.0
idna                    3.3
importlib-metadata      4.6.4
incremental             21.3.0
jax                     0.4.38
jaxlib                  0.4.38
jeepney                 0.7.1
keyring                 23.5.0
kiwisolver              1.4.7
language-selector       0.1
launchpadlib            1.10.16
lazr.restfulclient      0.14.4
lazr.uri                1.0.6
libtorrent              2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
louis                   3.20.0
macaroonbakery          1.3.1
Mako                    1.1.3
MarkupSafe              2.0.1
matplotlib              3.10.0
mediapipe               0.10.20
ml_dtypes               0.5.0
more-itertools          8.10.0
netaddr                 0.8.0
netifaces               0.11.0
numpy                   1.26.4
oauthlib                3.2.0
olefile                 0.46
opencv-contrib-python   4.10.0.84
opencv-python           4.10.0.84
opt_einsum              3.4.0
packaging               24.2
pandas                  2.2.3
pillow                  11.0.0
pip                     22.0.2
plotly                  5.24.1
pop-transition          1.1.2
protobuf                4.25.5
pyarrow                 18.1.0
pyasn1                  0.4.8
pyasn1-modules          0.2.1
pycairo                 1.20.1
pycparser               2.22
pycups                  2.0.1
pydbus                  0.6.0
pygame                  2.1.2
PyGObject               3.42.1
PyHamcrest              2.0.2
PyJWT                   2.3.0
pymacaroons             0.13.0
PyNaCl                  1.5.0
pyOpenSSL               21.0.0
pyparsing               2.4.7
pyRFC3339               1.1
python-apt              2.4.0+ubuntu3
python-dateutil         2.9.0.post0
python-debian           0.1.43+ubuntu1.1
python-gnupg            0.4.8
python-xlib             0.29
pytz                    2022.1
pyxdg                   0.27
PyYAML                  5.4.1
rencode                 1.0.6
repolib                 2.2.1
repoman                 1.4.0
requests                2.25.1
scipy                   1.14.1
screen-resolution-extra 0.0.0
SecretStorage           3.3.1
sentencepiece           0.2.0
service-identity        18.1.0
sessioninstaller        0.0.0
setproctitle            1.2.2
setuptools              59.6.0
six                     1.16.0
sounddevice             0.5.1
systemd-python          234
tenacity                9.0.0
Twisted                 22.1.0
tzdata                  2024.2
ubuntu-drivers-common   0.0.0
ufw                     0.36.1
urllib3                 1.26.5
wadllib                 1.3.6
wheel                   0.37.1
xdg                     5
xkit                    0.0.0
zipp                    1.0.0
zope.interface          5.4.0
